% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndiaFoodcrop.R
\name{calcIndiaFoodcrop}
\alias{calcIndiaFoodcrop}
\title{calcIndiaFoodcrop
Calculates foodcrops data for India taken from here: https://eands.dacnet.nic.in/APY_96_To_07.htm}
\usage{
calcIndiaFoodcrop(subtype = "Area")
}
\arguments{
\item{subtype}{Area, Yield, or Production}
}
\value{
magpie object containing Area, Yield, and Production data.
}
\description{
calcIndiaFoodcrop
Calculates foodcrops data for India taken from here: https://eands.dacnet.nic.in/APY_96_To_07.htm
}
\examples{
\dontrun{
a <- madrat::calcOutput("IndiaFoodcrop", subtype = "Area")
}
}
\author{
Anastasis Giannousakis
}
