% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndiaStateFoodcrop.R
\name{calcIndiaStateFoodcrop}
\alias{calcIndiaStateFoodcrop}
\title{calcIndiaStateFoodcrop
Calculates foodcrops data for India downloaded from:
 UPAg - Unified Portal for Agricultural Statistics by  Department of Agriculture & Farmers Welfare}
\usage{
calcIndiaStateFoodcrop(subtype = "Area")
}
\arguments{
\item{subtype}{Area, Yield, or Production}
}
\value{
magpie object containing Area, Yield, and Production data.
}
\description{
calcIndiaStateFoodcrop
Calculates foodcrops data for India downloaded from:
 UPAg - Unified Portal for Agricultural Statistics by  Department of Agriculture & Farmers Welfare
}
\examples{
\dontrun{
a <- madrat::calcOutput("IndiaFoodcrop", subtype = "Area")
}
}
\author{
Ankit Saha
}
