% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctIndiaAPY.R
\name{correctIndiaAPY}
\alias{correctIndiaAPY}
\title{correctIndiaAPY}
\usage{
correctIndiaAPY(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie objects with corrected IndiaAPY data
}
\description{
Correct IndiaAPY data
}
\examples{
\dontrun{
readSource("IndiaAPY", convert = "onlycorrect")
}

}
\author{
Anastasis Giannousakis
}
