% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctIndiaStateAPY.R
\name{correctIndiaStateAPY}
\alias{correctIndiaStateAPY}
\title{correctIndiaStateAPY}
\usage{
correctIndiaStateAPY(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie objects with corrected IndiaStateAPY data
}
\description{
Correct IndiaStateAPY data
}
\examples{
\dontrun{
readSource("IndiaStateAPY", convert = "onlycorrect")
}

}
\author{
Ankit Saha
}
