% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIndiaAPY.R
\name{downloadIndiaAPY}
\alias{downloadIndiaAPY}
\title{downloadIndiaAPY}
\usage{
downloadIndiaAPY()
}
\description{
Function to download data from https://eands.dacnet.nic.in/APY_96_To_07.htm
}
\author{
Anastasis Giannousakis
}
