% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullDATAINDIA.R
\name{fullDATAINDIA}
\alias{fullDATAINDIA}
\title{fullDATAINDIA
Calculates foodcrops data for India taken from here: https://eands.dacnet.nic.in/APY_96_To_07.htm}
\usage{
fullDATAINDIA()
}
\value{
Foodcrop India data
}
\description{
fullDATAINDIA
Calculates foodcrops data for India taken from here: https://eands.dacnet.nic.in/APY_96_To_07.htm
}
\examples{
\dontrun{
a <- retrieveData("DATAINDIA")
}
}
\author{
Anastasis Giannousakis
}
