% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfable-package.R
\docType{package}
\name{mrfable-package}
\alias{mrfable-package}
\alias{_PACKAGE}
\alias{mrfable}
\title{The mrfable R package (runs with madrat)}
\description{
Provides functions to read in FABLE project data for India from here: https://eands.dacnet.nic.in/APY_96_To_07.htm
}
\author{
Anastasis Giannousakis
}
