% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIndiaAPY.R
\name{readIndiaAPY}
\alias{readIndiaAPY}
\title{readIndiaAPY
This function reads foodcrops data for India taken from here: https://eands.dacnet.nic.in/APY_96_To_07.htm}
\usage{
readIndiaAPY(subtype = NA)
}
\arguments{
\item{subtype}{Area, Yield, or Production}
}
\value{
magpie object containing Area, Yield, and Production data.
}
\description{
readIndiaAPY
This function reads foodcrops data for India taken from here: https://eands.dacnet.nic.in/APY_96_To_07.htm
}
\examples{
\dontrun{ a <- madrat::readSource(type="IndiaAPY",subtype="Rice",convert="onlycorrect") }
}
\author{
Anastasis Giannousakis
}
