% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIndiaStateAPY.R
\name{readIndiaStateAPY}
\alias{readIndiaStateAPY}
\title{readIndiaStateAPY}
\usage{
readIndiaStateAPY()
}
\value{
magpie object containing Area, Yield, and Production data.
}
\description{
This function reads foodcrops data for India downloaded from:
 UPAg - Unified Portal for Agricultural Statistics by  Department of Agriculture & Farmers Welfare
 https://upag.gov.in/dash-reports/stateprofile?rtab=State+Profile%3A+Crop-wise+APY&rtype=reports
}
\examples{
\dontrun{ a <- madrat::readSource(type="IndiaStateAPY",subtype="Rice",convert="onlycorrect") }
}
\author{
Ankit Saha
}
