% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndiaStateFoodcrop.R
\name{calcIndiaStateFoodcrop}
\alias{calcIndiaStateFoodcrop}
\title{calcIndiaStateFoodcrop}
\usage{
calcIndiaStateFoodcrop(subtype = "Area")
}
\arguments{
\item{subtype}{Area, Yield, or Production}
}
\value{
magpie object containing Area, Yield, and Production data.
}
\description{
Calculates foodcrops data for India downloaded from:
 UPAg - Unified Portal for Agricultural Statistics by  Department of Agriculture & Farmers Welfare
 https://upag.gov.in/dash-reports/stateprofile?rtab=State+Profile%3A+Crop-wise+APY&rtype=reports
}
\examples{
\dontrun{
a <- madrat::calcOutput("IndiaStateFoodcrop", subtype = "Area")
}
}
\author{
Ankit Saha
}
