% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIndiaStateAPY.R
\name{readIndiaStateAPY}
\alias{readIndiaStateAPY}
\title{readIndiaStateAPY
This function reads foodcrops data for India downloaded from:
 UPAg - Unified Portal for Agricultural Statistics by  Department of Agriculture & Farmers Welfare}
\usage{
readIndiaStateAPY()
}
\value{
magpie object containing Area, Yield, and Production data.
}
\description{
readIndiaStateAPY
This function reads foodcrops data for India downloaded from:
 UPAg - Unified Portal for Agricultural Statistics by  Department of Agriculture & Farmers Welfare
}
\examples{
\dontrun{ a <- madrat::readSource(type="IndiaAPY",subtype="Rice",convert="onlycorrect") }
}
\author{
Ankit Saha
}
